requirements.txt
beautifulsoup4==4.9.3
dnspython==1.16.0
pymongo==3.11.2
python-dotenv==0.15.0
requests==2.25.1
sentry-sdk==0.19.5
soupsieve==2.1